Remove "region" division
The `region` field used in SISVIG is a non-official regional
classification, essentially for the province of Panama:

    Panamá:
      - Panamá Metro
      - Panamá Oeste
      - Panamá Este
      - San Miguelito
      - Panamá Norte

Demote our indigenous region subdivision to a field inside the district
(indicating that the district belongs to this region), as this is not
part of the "official" address format.

Handle the form submission via a shim.  Evidently, this pre-dates the
creation of the Panamá Oeste province.  The shim should not include this
option.  It is still unclear to what region a given corregimiento
belongs to (some of them).