Fix typo
Walk-in -> Outpatient.

The term in Spanish is correct, so, no visible changes.

Walk-in: A previously unscheduled walk-in visit. [Appointment]

Outpatient: (Ambulatory) A comprehensive term for health care provided
in a healthcare facility (e.g. a practitioner's office, clinic setting,
or hospital) on a nonresident basis. The term ambulatory usually implies
that the patient has come to the location and is not assigned to a bed.
Sometimes referred to as an outpatient encounter. [Encounter]