Fix age duration calculation
The previous calculation was still imprecise.

The new method prefers a simpler approach, calculating an age the
way you learn in school.

While here, add a guard against negative durations.

Do not calculate a pediatric age based upon an age, as the age in weeks
can be imprecise for newborns.

Fix the tests accordingly.