pdf: Fix the display of notes at end of page
There was a bug at the end of the page: if there was no space to display
the notes, it produced a couple of empty pages, and bogus notes.

Draw notes as a table, and place it inside the last row of each
department.

As a consequence, the notes are printed with the department title, if
necessary.