Add a cedula validator
Validate cedulas and NT RUCs.  All other RUCs are not yet validated.

Note that the notation for AV/PI has been kept as usual, e.g., 8AV-1-1.
The DGI uses the NT format: 8-AV-1-1.  We might need to adapt it in the
future for electronic invoicing purposes.

The DGI seems to allow the "asiento/ficha" to contain more than 5/6
characters (9 actually).  This might have to be adapted as well.

Disabled for now until SVI access is granted.